Break out Com2* subclasses
