+++ src/System.Windows.Forms/src/System/Windows/Forms/ComponentModel/COM2Interop/COM2AboutBoxPropertyDescriptor.cs
@@ -5,16 +5,14 @@
 #nullable disable
 
 using System.ComponentModel;
-using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Drawing.Design;
 using System.Runtime.InteropServices;
-using Windows.Win32.System.Com;
 using static Interop;
 
 namespace System.Windows.Forms.ComponentModel.Com2Interop
 {
-    internal class Com2AboutBoxPropertyDescriptor : Com2PropertyDescriptor
+    internal partial class Com2AboutBoxPropertyDescriptor : Com2PropertyDescriptor
     {
         private TypeConverter _converter;
         private UITypeEditor _editor;
@@ -66,37 +64,5 @@ namespace System.Windows.Forms.ComponentModel.Com2Interop
         public override void SetValue(object component, object value) => throw new ArgumentException();
 
         public override bool ShouldSerializeValue(object component) => false;
-
-        public class AboutBoxUITypeEditor : UITypeEditor
-        {
-            public override unsafe object EditValue(ITypeDescriptorContext context, IServiceProvider provider, object value)
-            {
-                object component = context.Instance;
-
-                if (Marshal.IsComObject(component) && component is Oleaut32.IDispatch pDisp)
-                {
-                    EXCEPINFO pExcepInfo = new();
-                    DISPPARAMS dispParams = new();
-                    Guid g = Guid.Empty;
-                    HRESULT hr = pDisp.Invoke(
-                        Ole32.DispatchID.ABOUTBOX,
-                        &g,
-                        PInvoke.GetThreadLocale(),
-                        DISPATCH_FLAGS.DISPATCH_METHOD,
-                        &dispParams,
-                        null,
-                        &pExcepInfo,
-                        null);
-                    Debug.Assert(hr.Succeeded, "Failed to launch about box.");
-                }
-
-                return value;
-            }
-
-            public override UITypeEditorEditStyle GetEditStyle(ITypeDescriptorContext context)
-            {
-                return UITypeEditorEditStyle.Modal;
-            }
-        }
     }
 }
