+++ src/System.Windows.Forms/src/System/Windows/Forms/ComponentModel/COM2Interop/Com2Variant.cs
@@ -1,0 +1,13 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Windows.Forms.ComponentModel.Com2Interop
+{
+    /// <summary>
+    ///  A class included so we can recognize a variant properly.
+    /// </summary>
+    public class Com2Variant
+    {
+    }
+}
