+++ src/System.Windows.Forms/src/System/Windows/Forms/ComponentModel/COM2Interop/ComNativeDescriptor.cs
@@ -12,14 +12,13 @@ using System.Runtime.InteropServices;
 using Windows.Win32.System.Com;
 using static Interop;
 using static Interop.Ole32;
-using static System.TrimmingConstants;
 
 namespace System.Windows.Forms.ComponentModel.Com2Interop
 {
     /// <summary>
     ///  Top level mapping layer between COM Object and TypeDescriptor.
     /// </summary>
-    internal class ComNativeDescriptor : TypeDescriptionProvider
+    internal partial class ComNativeDescriptor : TypeDescriptionProvider
     {
         private static ComNativeDescriptor handler;
 
@@ -467,127 +466,6 @@ namespace System.Windows.Forms.ComponentModel.Com2Interop
                 {
                     currentEditor = subEditor;
                 }
-            }
-        }
-
-        /// <summary>
-        ///  This type descriptor sits on top of a ComNativeDescriptor
-        /// </summary>
-        private sealed class ComTypeDescriptor : ICustomTypeDescriptor
-        {
-            private readonly ComNativeDescriptor _handler;
-            private readonly object _instance;
-
-            /// <summary>
-            ///  Creates a new WalkingTypeDescriptor.
-            /// </summary>
-            internal ComTypeDescriptor(ComNativeDescriptor handler, object instance)
-            {
-                _handler = handler;
-                _instance = instance;
-            }
-
-            /// <summary>
-            ///  ICustomTypeDescriptor implementation.
-            /// </summary>
-            AttributeCollection ICustomTypeDescriptor.GetAttributes()
-            {
-                return _handler.GetAttributes(_instance);
-            }
-
-            /// <summary>
-            ///  ICustomTypeDescriptor implementation.
-            /// </summary>
-            string ICustomTypeDescriptor.GetClassName()
-            {
-                return GetClassName(_instance);
-            }
-
-            /// <summary>
-            ///  ICustomTypeDescriptor implementation.
-            /// </summary>
-            string ICustomTypeDescriptor.GetComponentName()
-            {
-                return GetName(_instance);
-            }
-
-            /// <summary>
-            ///  ICustomTypeDescriptor implementation.
-            /// </summary>
-            [RequiresUnreferencedCode(AttributesRequiresUnreferencedCodeMessage)]
-            TypeConverter ICustomTypeDescriptor.GetConverter()
-            {
-                return GetConverter();
-            }
-
-            /// <summary>
-            ///  ICustomTypeDescriptor implementation.
-            /// </summary>
-            [RequiresUnreferencedCode(EventDescriptorRequiresUnreferencedCodeMessage)]
-            EventDescriptor ICustomTypeDescriptor.GetDefaultEvent()
-            {
-                return GetDefaultEvent();
-            }
-
-            /// <summary>
-            ///  ICustomTypeDescriptor implementation.
-            /// </summary>
-            [RequiresUnreferencedCode(PropertyDescriptorPropertyTypeMessage)]
-            PropertyDescriptor ICustomTypeDescriptor.GetDefaultProperty()
-            {
-                return _handler.GetDefaultProperty(_instance);
-            }
-
-            /// <summary>
-            ///  ICustomTypeDescriptor implementation.
-            /// </summary>
-            [RequiresUnreferencedCode(EditorRequiresUnreferencedCode)]
-            object ICustomTypeDescriptor.GetEditor(Type editorBaseType)
-            {
-                return GetEditor(_instance, editorBaseType);
-            }
-
-            /// <summary>
-            ///  ICustomTypeDescriptor implementation.
-            /// </summary>
-            EventDescriptorCollection ICustomTypeDescriptor.GetEvents()
-            {
-                return GetEvents();
-            }
-
-            /// <summary>
-            ///  ICustomTypeDescriptor implementation.
-            /// </summary>
-            [RequiresUnreferencedCode(FilterRequiresUnreferencedCodeMessage)]
-            EventDescriptorCollection ICustomTypeDescriptor.GetEvents(Attribute[] attributes)
-            {
-                return GetEvents();
-            }
-
-            /// <summary>
-            ///  ICustomTypeDescriptor implementation.
-            /// </summary>
-            [RequiresUnreferencedCode(PropertyDescriptorPropertyTypeMessage)]
-            PropertyDescriptorCollection ICustomTypeDescriptor.GetProperties()
-            {
-                return _handler.GetProperties(_instance);
-            }
-
-            /// <summary>
-            ///  ICustomTypeDescriptor implementation.
-            /// </summary>
-            [RequiresUnreferencedCode($"{PropertyDescriptorPropertyTypeMessage} {FilterRequiresUnreferencedCodeMessage}")]
-            PropertyDescriptorCollection ICustomTypeDescriptor.GetProperties(Attribute[] attributes)
-            {
-                return _handler.GetProperties(_instance);
-            }
-
-            /// <summary>
-            ///  ICustomTypeDescriptor implementation.
-            /// </summary>
-            object ICustomTypeDescriptor.GetPropertyOwner(PropertyDescriptor pd)
-            {
-                return _instance;
             }
         }
     }
